used longs instead of ints

+++ helloasm/helloworld.c
@@ -1,8 +1,9 @@
 #include <stdio.h>
 
-extern int sumfunc (int a, int b); // adds 2 numbers
-extern int powfunc (int a, int b); // elevates a to the b power
-extern void add10func (int *a); // increments variable a in 10
+//We use longs (8 bytes) and not ints (4 bytes) because the processor is 64 bit
+extern long sumfunc (long a, long b); // adds 2 numbers
+extern long powfunc (long a, long b); // elevates a to the b power
+extern void add10func (long *a); // increments variable a in 10
 
 int main() {
 	printf("Hello, World!\n");
@@ -10,7 +11,7 @@ int main() {
 	printf("sumfunc(4,3)=%i\n", sumfunc(4, 3));
 	printf("powfunc(3,3)=%i\n", powfunc(3, 3));
 
-    int mynumber=10; 
+    long mynumber=10; 
     printf ("mynumber before add10func: %i\n", mynumber);
     add10func (&mynumber);
     printf ("mynumber after add10func: %i\n", mynumber);
